requirements.txt
picamera
colour-science
flask
matplotlib
numpy
Pillow